requirements.txt
google-cloud==0.34.0
google-cloud-storage==2.0.0
google-api-python-client==2.42.0
google-auth==2.3.3